Separate code for collecting form data
